requirements_branchbot.txt
requests==2.32.3
beautifulsoup4==4.12.3
python-dotenv==1.0.1
invoke==2.2.0
pytest==8.3.2
gspread==6.1.2
google-auth==2.35.0